add more kavanagh

+++ scripts/send-poem.py
@@ -58,9 +58,10 @@ if args.write_to_repo:
 
         history.loc[index, "date"] = date
         history.loc[index, "time"] = time
-        history.loc[index, "timestamp"] = int(now.timestamp)
+        history.loc[index, "timestamp"] = now.timestamp
         history.loc[index, "title"] = p.key
         history.loc[index, "author"] = p.author.key
+        history["timestamp"] = history["timestamp"].astype(int)
 
     daily_poems = {}
     for index, entry in history.iterrows():
